+++ src/main.jsx
@@ -14,6 +14,7 @@ import StockPurchase from './pages/StockPurchase'
 import Payments from './pages/Payments'
 import Ledger from './pages/Ledger'
 import Settings from './pages/Settings'
+import OtherExpenses from './pages/OtherExpenses'
 import { ipc } from './lib/ipc'
 
 function App() {
@@ -56,6 +57,7 @@ function App() {
           case 't': navigate('/stock-purchase'); break
           case 'w': navigate('/payments'); break
           case 'l': navigate('/ledger'); break
+          case 'o': navigate('/other-expenses'); break
           case 'r': navigate('/'); break
           case 's': navigate('/settings'); break
         }
@@ -83,6 +85,7 @@ function App() {
             <Route path="/payments" element={<Payments />} />
             <Route path="/ledger" element={<Ledger />} />
             <Route path="/settings" element={<Settings />} />
+            <Route path="/other-expenses" element={<OtherExpenses />} />
           </Routes>
         </main>
       </div>